Refactors time parameter types for consistency
Ensures consistent use of `uint32_t` for time-related parameters within the operation manager.

This change addresses potential inconsistencies and improves code readability by using a single type for representing time values throughout the relevant functions. It also fixes format string usage to match the data type.

+++ PAC/common/operation_mngr.h
@@ -767,7 +767,7 @@ class operation
         ///
         /// @param new_step - номер шага (с единицы).
         /// @param cooperative_time - время совместной работы (сек).
-        void to_step( unsigned int new_step, unsigned long cooperative_time = 0 );
+        void to_step( unsigned int new_step, uint32_t cooperative_time = 0 );
 
         /// @brief Переход к следующему шагу.
         ///
